services/user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "User-service"
authors = ["Barashikhin Egor"]
readme = "README.md"
packages = [{include = "user_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
pydantic-settings = "^2.0.3"
alembic = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
